refactor(walktrip): move settings out of walktrip
