+++ src/components/WalkTrip/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import classNames from 'classnames'
-import { useSettings } from 'settings/SettingsProvider'
 import { useWalkTrip } from 'hooks/use-walk-trip/useWalkTrip'
 import { Coordinates } from 'src/types'
 import { formatWalkTrip } from 'utils/formatting'
@@ -9,14 +8,15 @@ import classes from './WalkTrip.module.scss'
 function WalkTrip({
     className,
     coordinates,
+    hideWalkInfo,
 }: {
     className?: string
     coordinates: Coordinates
+    hideWalkInfo: boolean
 }) {
-    const [settings] = useSettings()
     const { walkTrip } = useWalkTrip(coordinates)
 
-    if (settings.hideWalkInfo || !walkTrip) return null
+    if (hideWalkInfo || !walkTrip) return null
 
     return (
         <div className={classNames(classes.WalkTrip, className)}>
