+++ src/routes/Dashboard/Chrono/ChronoDepartureTile/ChronoDepartureTile.tsx
@@ -10,8 +10,8 @@ import {
     filterHidden,
     toDeparture,
 } from 'hooks/use-stop-place-with-estimated-calls/departure'
-import { WalkTrip } from 'components/WalkTrip/WalkTrip'
 import { Loader } from 'components/Loader'
+import { WalkTrip } from 'components/WalkTrip'
 import { ErrorTile } from 'tiles/dashboard/ErrorTile'
 import { EmptyStopTile } from 'tiles/dashboard/EmptyStopTile'
 import { HeaderCell, Table, TableHead, TableRow } from '@entur/table'
@@ -73,6 +73,7 @@ function ChronoDepartureTile({ stopPlaceId }: { stopPlaceId: string }) {
                     longitude: stopPlaceWithEstimatedCalls.longitude,
                     latitude: stopPlaceWithEstimatedCalls.latitude,
                 }}
+                hideWalkInfo={settings.hideWalkInfo}
             />
             <Table fixed>
                 <TableHead className={classes.TableHead}>
