--- src/scenarios/Map/components/StopPlaceMarkers.tsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import { useStopPlaceIds } from 'hooks/use-stop-place-ids/useStopPlaceIds'
-import { StopPlaceMarker } from './StopPlaceMarker'
-
-function StopPlaceMarkers() {
-    const { stopPlaceIds } = useStopPlaceIds()
-
-    return (
-        <>
-            {stopPlaceIds.map((stopPlaceId) => (
-                <StopPlaceMarker stopPlaceId={stopPlaceId} key={stopPlaceId} />
-            ))}
-        </>
-    )
-}
-
-export { StopPlaceMarkers }
